Fix date format and sidebar styling
Updated date and time format to "May 22, 2025, 06:28 PM" with three-letter month abbreviations in Deposits & Transactions, Token & Package Control, and Users & Wallets components. Made the sidebar non-scrollable and fixed its position.
Files edited: `src/components/admin/DepositsTransactions.tsx`, `src/components/admin/TokenPackageControl.tsx`, `src/components/admin/UsersWallets.tsx`, `src/components/ui/sidebar.tsx`.

+++ src/components/admin/DepositsTransactions.tsx
@@ -1,4 +1,3 @@
-
 import { useState } from 'react';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { Button } from '@/components/ui/button';
@@ -46,7 +45,7 @@ export function DepositsTransactions() {
     const date = new Date(dateString);
     return date.toLocaleDateString('en-US', {
       year: 'numeric',
-      month: 'long',
+      month: 'short',
       day: 'numeric',
       hour: 'numeric',
       minute: '2-digit',
